Added BaseMenuActivity with the overriden methods for all Activities that need the AboutUs Menu
