+++ app/src/main/java/ch/noseryoung/lernendeverwaltung/activity/BaseMenuActivity.java
@@ -1,0 +1,33 @@
+package ch.noseryoung.lernendeverwaltung.activity;
+
+import android.content.Intent;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import ch.noseryoung.lernendeverwaltung.R;
+
+public class BaseMenuActivity extends AppCompatActivity {
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+
+        MenuInflater menuInflater = getMenuInflater();
+        menuInflater.inflate(R.menu.main_menu,menu);
+
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        super.onOptionsItemSelected(item);
+
+        if(item.getItemId() == R.id.aboutus) {
+            Intent intend = new Intent(this, AboutUsActivity.class);
+            startActivity(intend);
+            return true;
+        }
+        return false;
+    }
+}
