+++ app/src/main/java/ch/noseryoung/lernendeverwaltung/activity/FullScreenImageActivity.java
@@ -1,14 +1,13 @@
 package ch.noseryoung.lernendeverwaltung.activity;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.widget.ImageView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import ch.noseryoung.lernendeverwaltung.R;
 import ch.noseryoung.lernendeverwaltung.manager.UserImageViewManager;
@@ -33,8 +32,8 @@ public class FullScreenImageActivity extends AppCompatActivity {
 
         if (callingActivityIntent != null) {
 
-            if (callingActivityIntent.hasExtra(UserlistActivity.EXTRA_USER_ID)) {
-                int apprenticeID = callingActivityIntent.getIntExtra(UserlistActivity.EXTRA_USER_ID, 0);
+            if (callingActivityIntent.hasExtra(UserlistMenuActivity.EXTRA_USER_ID)) {
+                int apprenticeID = callingActivityIntent.getIntExtra(UserlistMenuActivity.EXTRA_USER_ID, 0);
                 setFullscreenImageById(apprenticeID);
             } else{
                 Uri apprenticePhotoUri = callingActivityIntent.getData();
