+++ app/src/main/res/layout/activity_new_user.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/default_background"
-    tools:context=".activity.NewUserActivity">
+    tools:context=".activity.NewUserMenuActivity">
 
     <Button
         android:id="@+id/newUser_createButton"
