+++ app/src/main/res/layout/activity_user_data.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/default_background"
-    tools:context=".activity.UserdataActivity">
+    tools:context=".activity.UserdataMenuActivity">
 
     <TextView
         android:id="@+id/userdata_firstname"
