+++ app/src/main/res/layout/activity_userlist.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/default_background"
-    tools:context=".activity.UserlistActivity">
+    tools:context=".activity.UserlistMenuActivity">
 
     <Button
         android:id="@+id/userlist_createApprenticeButton"
